Add mapping for RDF image
